Tests\Unit: updated to support new signatures KP-8034

+++ src/Tests/WB.Tests.Unit/SharedKernels/DataCollection/InterviewTests/when_answer_with_201_selected_options_on_multioption_question_which_triggers_roster_and_has_maxAllowedAnswers.cs
@@ -16,7 +16,7 @@ namespace WB.Tests.Unit.SharedKernels.DataCollection.InterviewTests
 
             rosterSizeQuestionId = Guid.Parse("33333333333333333333333333333333");
 
-            answers = new decimal[201];
+            answers = new int[201];
             for (int i = 0; i < answers.Length; i++)
             {
                 answers[i] = i;
@@ -45,7 +45,7 @@ namespace WB.Tests.Unit.SharedKernels.DataCollection.InterviewTests
         private static Interview interview;
         private static Guid userId;
         private static Guid rosterSizeQuestionId;
-        private static decimal[] answers;
+        private static int[] answers;
         private static InterviewException exception;
 
     }
