+++ src/Tests/WB.Tests.Unit/SharedKernels/DataCollection/InterviewTests/when_answering_real_question_which_is_roster_title_for_2_rosters_and_roster_level_is_1_and_answer_is_1_with_a_half.cs
@@ -45,7 +45,7 @@ namespace WB.Tests.Unit.SharedKernels.DataCollection.InterviewTests
         };
 
         Because of = () =>
-            interview.AnswerNumericRealQuestion(userId, questionId, rosterVector, DateTime.Now, (decimal) 1.5);
+            interview.AnswerNumericRealQuestion(userId, questionId, rosterVector, DateTime.Now, 1.5);
 
         Cleanup stuff = () =>
         {
